ai-service/requirements.txt
# Python dependencies for CV Analyzer AI Service
# Install with: pip install -r requirements.txt

# ============================================================================
# Core Framework
# ============================================================================
# FastAPI - Modern, fast web framework
fastapi==0.115.5

# Uvicorn - ASGI server
uvicorn[standard]==0.32.1

# ============================================================================
# Microsoft Agent Framework (PREVIEW)
# ============================================================================
# IMPORTANT: Use --pre flag during installation
# pip install agent-framework-azure-ai --pre
agent-framework-azure-ai>=0.1.0

# ============================================================================
# Azure SDK
# ============================================================================
# Azure Identity - Authentication
azure-identity==1.19.0

# Azure AI Inference (may be needed for Agent Framework)
azure-ai-inference==1.0.0b7

# ============================================================================
# Data Validation & Serialization
# ============================================================================
# Pydantic - Data validation
pydantic==2.10.3

# Pydantic Settings - Configuration management
pydantic-settings==2.6.1

# ============================================================================
# Utilities
# ============================================================================
# Python dotenv - Environment variable management
python-dotenv==1.0.1

# HTTP Client (if needed for additional integrations)
httpx==0.28.1

# ============================================================================
# Development Dependencies (optional)
# ============================================================================
# Pytest - Testing
pytest==8.3.4
pytest-asyncio==0.24.0

# Black - Code formatting
black==24.10.0

# Ruff - Linting
ruff==0.8.4

# MyPy - Type checking
mypy==1.13.0
